Sorry I Changed A lot Tell me What you Think or if there is a better way to break this down
