+++ WordGuess/GitSetupList.txt
@@ -1,3 +1,2 @@
 People who have setup GitHub so far:
 Daniel
-null
