+++ WordGuess/Read Me
@@ -1,0 +1,4 @@
+READ ME
+
+Main.java is to configure the git set up
+UserName.java is for the GUI interface
