+++ WordGuess/src/GUI/InterfacePanels/Toolbar.java
@@ -1,4 +1,4 @@
-package GUI;
+package GUI.InterfacePanels;
 
 import java.awt.FlowLayout;
