+++ WordGuess/src/GUI/MainMenuBar/GameMenuTab.java
@@ -1,0 +1,6 @@
+package GUI.MainMenuBar;
+
+public class GameMenuTab 
+{
+
+}
