+++ WordGuess/src/GUI/MainMenuBar/MainMenuBar.java
@@ -1,0 +1,6 @@
+package GUI.MainMenuBar;
+
+public class MainMenuBar 
+{
+
+}
