+++ WordGuess/src/GUI/MainMenuBar/ShowMenuTab.java
@@ -1,0 +1,5 @@
+package GUI.MainMenuBar;
+
+public class ShowMenuTab {
+
+}
