+++ WordGuess/src/GUI/MainMenuBar/WindowMenuTab.java
@@ -1,0 +1,5 @@
+package GUI.MainMenuBar;
+
+public class WindowMenuTab {
+
+}
